requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.1
anyio==4.11.0
asteroid-filterbanks==0.4.0
attrs==25.4.0
audioread==3.1.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
colorlog==6.10.1
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
einops==0.8.1
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.2
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
joblib==1.5.2
julius==0.2.7
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
lightning==2.5.6
lightning-utilities==0.15.2
llvmlite==0.45.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
networkx==3.5
numba==0.62.1
numpy==2.3.4
openai-whisper==20250625
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
optuna==4.5.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
platformdirs==4.5.0
pooch==1.8.2
primePy==1.3
propcache==0.4.1
protobuf==6.33.0
pyannote-audio==4.0.1
pyannote-core==6.0.1
pyannote-database==6.1.0
pyannote-metrics==4.0.0
pyannote-pipeline==4.0.0
pyannoteai-sdk==0.3.0
pycparser==2.23
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytorch-lightning==2.5.6
pytorch-metric-learning==2.9.0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rich==14.2.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==1.0.0
SQLAlchemy==2.0.44
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.12.0
torch==2.9.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.9.0
torchcodec==0.8.1
torchmetrics==1.8.2
tqdm==4.67.1
typer-slim==0.20.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
yarl==1.22.0
zipp==3.23.0
